뭐 많이 수정함

+++ cloudFunding/src/main/webapp/WEB-INF/views/project/community/updatePostForm.jsp
@@ -209,9 +209,11 @@
 	});
 	
 	function deletePost(){
-		var postCode = $("input[name=postCode]").val();		
-		var projectCode = $("input[name=projectCode]").val();		
-		location.href="deletePost.do?postCode=" + postCode + "&projectCode=" + projectCode;
+		if(confirm("게시글을 삭제하시겠습니까? 삭제된 글을 복구할 수 없습니다.")) {
+			var postCode = $("input[name=postCode]").val();		
+			var projectCode = $("input[name=projectCode]").val();		
+			location.href="deletePost.do?postCode=" + postCode + "&projectCode=" + projectCode;
+		}
 	}
 	
 	function updatePost(){
